requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.5
colorama==0.4.6
cryptography==41.0.2
cssselect2==0.7.0
Django==4.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-windowsauth==1.4.0
django-xhtml2pdf==0.0.4
djangorestframework==3.14.0
html5lib==1.1
idna==3.4
ldap3==2.9.1
lxml==4.9.3
oscrypto==1.3.0
Pillow==10.0.0
pyasn1==0.5.0
pycparser==2.21
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
pypdf==3.12.1
pypng==0.20220715.0
python-bidi==0.4.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
schedule==1.2.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.3
webencodings==0.5.1
wfastcgi==3.0.0
xhtml2pdf==0.2.11
